Updated to use size_t to allocate the sintarray which can exceed to Gigs in size when using delocalized projectors on large problems.
git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@3919 eb38b3b0-33ea-0310-94ad-d43fea0d214f

+++ Headers/Kpoint.h
@@ -99,7 +99,7 @@ public:
     KpointType *newsint_local;
 
     // Size of the sint arrays
-    int sint_size;
+    size_t sint_size;
 
     // Pointers to nv, ns, and Bns. Each of these arrays is dimensioned (NL_BLOCK_SIZE, P0_BASIS).
     // since we apply the non-local operators in blocks for efficiency and to save memory
